refactor: 도메인 로직을 domain/ 으로 이동
미션 2단계의 '도메인 로직은 domain/ 하위로'에 따른 변경.
