+++ src/App.js
@@ -1,4 +1,4 @@
-const RacingGame = require('./RacingGame');
+const RacingGame = require('./domain/RacingGame');
 const InputView = require('./view/InputView');
 const OutputView = require('./view/OutputView');
